requirements.txt
amqp==2.3.2
argh==0.26.2
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
gevent==1.3.7
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
pathtools==0.1.2
pycparser==2.19
pytz==2018.6
PyYAML==3.13
redis==2.10.6
schedule==0.6.0
shadowsocks-py==2.9.1
six==1.11.0
SQLAlchemy==1.2.12
vine==1.1.4
watchdog==0.9.0
Werkzeug==0.14.1
